Don't exit early if keys n/a in fu-mirror
It is common for an encrypted tarball with FreeBSD Update updates
to be available before the keys to decrypt it are published.  We
were tripping over this and exiting fu-mirror.sh with an error
rather than handling the condition correctly.  As a result, the
FreeBSD Update bits were not being marked as "up to date".

Revert 4957d242 which disabled the staleness check.